+++ http4k-format/jackson-yaml/src/test/kotlin/org/http4k/format/jacksonYamlTests.kt
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test
 class JacksonYamlBodyTest {
 
     @Test
-    fun `roundtrip list of arbitary objects to and from body`() {
+    fun `roundtrip list of arbitrary objects to and from body`() {
         val body = Body.auto<List<ArbObject>>().toLens()
 
         val obj = ArbObject("hello", ArbObject("world", null, listOf(1), true), emptyList(), false)
@@ -22,7 +22,7 @@ class JacksonYamlBodyTest {
     }
 
     @Test
-    fun `roundtrip array of arbitary objects to and from body`() {
+    fun `roundtrip array of arbitrary objects to and from body`() {
         val body = Body.auto<Array<ArbObject>>().toLens()
 
         val obj = ArbObject("hello", ArbObject("world", null, listOf(1), true), emptyList(), false)